hair_transfer/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.5
Pillow>=10.0.0
PyYAML>=6.0.1
tqdm>=4.66.1
scikit-image>=0.22.0
opencv-python>=4.7.0
face-alignment>=1.3.4
dlib-bin
ftfy
regex
clip @ git+https://github.com/openai/CLIP.git
gdown>=3.12.2
